+++ nameserver/entry.go
@@ -23,9 +23,8 @@ type Entry struct {
 	lHostname   string // lowercased (not exported, so not encoded by gob)
 	Version     int
 	Tombstone   int64 // timestamp of when it was deleted
-	// Denotes that the container is stopped; such entry is removed only
-	// when the container is destroyed or container does not exist after a restart of
-	// nameserver.
+	// Denotes that the container is stopped; the flag prevents a tombstoned entry
+	// from being deleted.
 	stopped bool
 }
