+++ nameserver/nameserver.go
@@ -72,6 +72,19 @@ func (s set) exist(item string) bool {
 	return ok
 }
 
+// Used to store pending events before nameserver has been initialized
+type eventType int
+
+const (
+	eventDestroyed eventType = iota
+	eventDied
+)
+
+type event struct {
+	eType eventType
+	ident string
+}
+
 func New(ourName mesh.PeerName, domain string, db db.DB,
 	isKnownPeer func(mesh.PeerName) bool) *Nameserver {
 
@@ -84,7 +97,6 @@ func New(ourName mesh.PeerName, domain string, db db.DB,
 	}
 }
 
-// NB: Can be called only before n.Start()
 func (n *Nameserver) SetGossip(gossip mesh.Gossip) {
 	n.gossip = gossip
 }
@@ -110,7 +122,7 @@ func (n *Nameserver) Start(existingContainerIDs []string) {
 }
 
 // Restores local entries from the database.
-// NB: called only during the initialization, thus the lock is not taken.
+// NB: called only during the initialization.
 func (n *Nameserver) restoreFromDB(existingContainerIDs []string) {
 	n.debugf("restore: starting...")
 	defer n.debugf("restore: finished.")
@@ -171,7 +183,6 @@ func (n *Nameserver) AddEntry(hostname, containerid string,
 	var found bool
 
 	n.Lock()
-
 	// Restore the stopped entries first
 	for i, e := range n.entries {
 		if e.Origin == origin && e.ContainerID == containerid && e.stopped {
@@ -185,17 +196,14 @@ func (n *Nameserver) AddEntry(hostname, containerid string,
 			entries = append(entries, e)
 		}
 	}
-
 	if !found {
 		entries = append(entries, n.entries.add(hostname, containerid, origin, addr))
 	}
-
 	n.snapshot()
 	n.Unlock()
 	n.broadcastEntries(entries...)
 }
 
-// TODO(mp) DRY
 func (n *Nameserver) RestoreEntries(containerid string) {
 	var entries Entries
 
@@ -211,7 +219,6 @@ func (n *Nameserver) RestoreEntries(containerid string) {
 	}
 	n.snapshot()
 	n.Unlock()
-
 	n.broadcastEntries(entries...)
 }
 
@@ -243,6 +250,80 @@ func (n *Nameserver) ReverseLookup(ip address.Address) (string, error) {
 	}
 	n.debugf("reverse lookup %s -> %s", ip, match.Hostname)
 	return match.Hostname, nil
+}
+
+func (n *Nameserver) ContainerStarted(ident string) {}
+
+func (n *Nameserver) ContainerDestroyed(ident string) {
+	n.Lock()
+	if !n.ready {
+		n.pendingEvents = append(n.pendingEvents, event{eventDestroyed, ident})
+		n.Unlock()
+		return
+	}
+	entries := n.containerDestroyed(ident)
+	n.Unlock()
+	n.broadcastEntries(entries...)
+}
+
+func (n *Nameserver) ContainerDied(ident string) {
+	n.Lock()
+	if !n.ready {
+		n.pendingEvents = append(n.pendingEvents, event{eventDied, ident})
+		n.Unlock()
+		return
+	}
+	entries := n.containerDied(ident)
+	n.Unlock()
+	n.broadcastEntries(entries...)
+}
+
+func (n *Nameserver) containerDestroyed(ident string) Entries {
+	entries := n.entries.forceTombstone(n.ourName, func(e *Entry) bool {
+		if e.ContainerID == ident {
+			n.infof("container %s destroyed; tombstoning entry %s", ident, e.String())
+			// Unset the flag to allow the entry be removed later on;
+			// Because the flag isn't exposed to other peers, we don't
+			// need to bump the version number if the entry has been previously
+			// tombstoned.
+			e.stopped = false
+			return true
+		}
+		return false
+	})
+	n.snapshot()
+	return entries
+}
+
+func (n *Nameserver) containerDied(ident string) Entries {
+	entries := n.entries.tombstone(n.ourName, func(e *Entry) bool {
+		if e.ContainerID == ident {
+			n.infof("container %s died; tombstoning entry %s", ident, e.String())
+			// We might want to restore the entry if the container comes back
+			e.stopped = true
+			return true
+		}
+		return false
+	})
+	n.snapshot()
+	return entries
+}
+
+// TODO(mp) replayEvents could merge all entries and do a single broadcast
+func (n *Nameserver) replayEvents() {
+	var entries Entries
+
+	for _, e := range n.pendingEvents {
+		switch e.eType {
+		case eventDestroyed:
+			entries = n.containerDestroyed(e.ident)
+		case eventDied:
+			entries = n.containerDied(e.ident)
+		}
+		n.broadcastEntries(entries...)
+	}
+
+	n.pendingEvents = nil
 }
 
 func (n *Nameserver) PeerGone(peer mesh.PeerName) {
